% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/locfdr.R
\name{AnalyticalOptim}
\alias{AnalyticalOptim}
\title{Optimize lambda0 and pi0, using Poisson version of analytical approach
in Efron's "Microarrays, Empirical Bayes, and the Two-Groups Model"

In the paper, he derives formulas for arbitrary support. We use support
[0, 1] for estimating the null. We initialize the search with lambda=0.05 and
 pi=0.95.}
\usage{
AnalyticalOptim(r0, r1, N)
}
\arguments{
\item{r0}{The number of 0's in the data.}

\item{r1}{The number of 1's in the data.}

\item{N}{The total number of points in the data.}
}
\description{
Optimize lambda0 and pi0, using Poisson version of analytical approach
in Efron's "Microarrays, Empirical Bayes, and the Two-Groups Model"

In the paper, he derives formulas for arbitrary support. We use support
[0, 1] for estimating the null. We initialize the search with lambda=0.05 and
 pi=0.95.
}

