% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/locfdr.R
\name{BuildTable}
\alias{BuildTable}
\title{Build a table corresponding to the histogram of data}
\usage{
BuildTable(data)
}
\arguments{
\item{data}{The raw data (before histogramming) to estimate a mixture
  density for.}
}
\value{
X A data frame whose two columns are integer and number of times that
   integer appears in the data.
}
\description{
Build a table corresponding to the histogram of data
}

