% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/locfdr.R
\name{LLConstructor}
\alias{LLConstructor}
\title{Return log-likelihood function giving Poisson version of equation (4.6)
in "Microarrays, Empirical Bayes, and the Two-Groups model" by Efron.}
\usage{
LLConstructor(r0, r1, N0, N)
}
\arguments{
\item{r0}{The number of points equal to 0}

\item{r1}{The number of points equal to 1}

\item{N0}{The number of points equal to either 0 or 1}

\item{N}{The total number of points}
}
\value{
LL A function of lambda0 and pi0, given the current data log likelihood
}
\description{
Return log-likelihood function giving Poisson version of equation (4.6)
in "Microarrays, Empirical Bayes, and the Two-Groups model" by Efron.
}

