% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/locfdr.R
\name{LocfdrFuns}
\alias{LocfdrFuns}
\title{Run local fdr functions}
\usage{
LocfdrFuns(data, df, knots)
}
\arguments{
\item{data}{The raw data (before histogramming) to estimate a mixture density
for.}

\item{df}{The number of degrees of freedom to use in the natural spline
in fitting the mixture density.}

\item{knots}{The positions of the knots to use in the natural spline fit.}
}
\value{
fdr The estimated local fdr function (can be called on numeric vectors).

pi0 The estimated proportion of data that lies in the poisson component

lambda0 The estimated parameter for the poisson component.

f0 The estimated null density function (can be called on numeric
   vectors.)

f The estimated mixture density function (can be called on numeric
   vectors.)
}
\description{
Run local fdr functions
}

