% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/locfdr.R
\name{MixtureDensity}
\alias{MixtureDensity}
\title{Compute full mixture density for data, assuming exponential family model}
\usage{
MixtureDensity(data, df, knots)
}
\arguments{
\item{data}{The raw data (before histogramming) to estimate a mixture density
for.}

\item{df}{The number of degrees of freedom to use in the natural spline
in fitting the mixture density.}

\item{knots}{The positions of the knots to use in the natural spline fit.}
}
\value{
f\_hat The estimated mixture density function (can be called on numeric
   vectors.)
}
\description{
Compute full mixture density for data, assuming exponential family model
}

