% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/locfdr.R
\name{NullDensity}
\alias{NullDensity}
\title{Estimate the density for the null data}
\usage{
NullDensity(data)
}
\arguments{
\item{data}{The raw data (before histogramming) to estimate a mixture density
  for.}
}
\value{
f_hat The estimated mixture density function (can be called on numeric
   vectors.)
}
\description{
Estimate the density for the null data
}

