% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/locfdr.R
\name{SummarizeLocfdr}
\alias{SummarizeLocfdr}
\title{Call and summarize output to LocfdrFuns}
\usage{
SummarizeLocfdr(x, df, knots)
}
\arguments{
\item{x}{The raw data (before histogramming) to estimate a mixture density for.}

\item{df}{The number of degrees of freedom to use in the natural spline
in fitting the mixture density.}

\item{knots}{The positions of the knots to use in the natural spline fit.}
}
\value{
pi0 The estimated proportion of data that lies in the poisson component

lambda0 The estimated parameter for the poisson component.

locfdr\_res The fdr, f0 and f functions evaluated on the support of x

locfdr\_fig A plot of the estimated locfdr fit
}
\description{
Call and summarize output to LocfdrFuns
}

